add phong model

+++ include/material.h
@@ -12,7 +12,7 @@ struct material {
   double k_s_r_ = 0; // 镜面反射系数 0 ～ 1
   double k_s_g_ = 0;
   double k_s_b_ = 0;
-  double n_ = 1; // 镜面高光指数
+  int n_ = 1; // 镜面高光指数
 
   static material material_sample();
 };
